Extract stub_genus_lookup
